Updated Deliverable
Updated src content of the deliverable

+++ TFTP/Deliverable/Package/Java/src/utils/Utility.java
@@ -2,7 +2,6 @@ package utils;
 
 import java.io.IOException;
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.math.RoundingMode;
 import java.net.DatagramPacket;
 import java.nio.ByteBuffer;
@@ -10,7 +9,6 @@ import java.nio.charset.StandardCharsets;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.Base64;
 
 /**
  * Contains all utility functions for used by other class files.
@@ -304,7 +302,7 @@ public class Utility {
 		return trimmedPacket;
     }
     
-    public double percentageValue(int subtotal, int total) {
+    public double percentageValue(int subtotal, long total) {
     	double raw_percent = (((double)subtotal/(double)total)*100);
     	BigDecimal bd = new BigDecimal(raw_percent).setScale(2, RoundingMode.HALF_UP);
     	return bd.doubleValue();
